upgraded to version 0.1.1
If the project has many entities with BehaviorTree, then the implementation and visibility of nodes can now be separated by interfaces for each individual entity. To do this, you need to create your own interface from IBehaviorTreeType. Then add the created interface to the controller of your entity and to the implementation of your nodes. In the BehaviorTree component, select the "Show Nodes By Type" option, and now when editing BehaviorTree of your entity, only those implemented with your interface will appear in the node search.